# docs/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SoFinder HTTP API",
    "version": "1.0",
    "description": "Stable same-origin authenticated API. Mutations also require X-CSRF-TOKEN."
  },
  "servers": [{"url": "/sofinder"}],
  "security": [{"sessionCookie": []}],
  "paths": {
...
    "/api/assets/search": {"get": {"operationId": "searchAssets", "parameters": [{"name": "q", "in": "query", "schema": {"type": "string", "maxLength": 200}}, {"name": "resources", "in": "query", "schema": {"type": "string", "description": "Comma-separated visible resource names"}}, {"name": "path", "in": "query", "schema": {"type": "string"}}, {"name": "fields", "in": "query", "schema": {"type": "string", "default": "name,title,alt,tags"}}, {"name": "tags", "in": "query", "schema": {"type": "string"}}, {"name": "extensions", "in": "query", "schema": {"type": "string"}}, {"name": "type", "in": "query", "schema": {"enum": ["all", "image", "document", "audio", "video", "archive", "other"]}}, {"name": "minSize", "in": "query", "schema": {"type": "integer", "minimum": 0}}, {"name": "maxSize", "in": "query", "schema": {"type": "integer", "minimum": 0}}, {"name": "modifiedAfter", "in": "query", "schema": {"type": "integer", "minimum": 0}}, {"name": "modifiedBefore", "in": "query", "schema": {"type": "integer", "minimum": 0}}, {"name": "offset", "in": "query", "schema": {"type": "integer", "minimum": 0, "default": 0}}, {"name": "limit", "in": "query", "schema": {"type": "integer", "minimum": 1, "maximum": 100, "default": 50}}], "responses": {"200": {"$ref": "#/components/responses/AssetSearch"}, "400": {"$ref": "#/components/responses/Error"}, "403": {"$ref": "#/components/responses/Error"}}}},
...
    "/api/assets/{id}": {"get": {"operationId": "getAsset", "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string", "pattern": "^[a-f0-9-]{36}$"}}], "responses": {"200": {"$ref": "#/components/responses/Asset"}, "404": {"$ref": "#/components/responses/Error"}}}},
...
  },
  "components": {
    "securitySchemes": {"sessionCookie": {"type": "apiKey", "in": "cookie", "name": "PHPSESSID"}},
...
    "schemas": {
...
      "ErrorEnvelope": {"$ref": "./schema/error-envelope.schema.json"},
...
      "AssetReference": {"$ref": "./schema/asset-reference.schema.json"},
...
      "Entry": {"type": "object", "additionalProperties": false, "required": ["path", "name", "directory", "size", "modifiedAt", "mimeType", "url", "capabilities"], "properties": {"path": {"type": "string"}, "name": {"type": "string"}, "directory": {"type": "boolean"}, "size": {"type": "integer", "minimum": 0}, "modifiedAt": {"type": "integer", "minimum": 0}, "mimeType": {"type": ["string", "null"]}, "url": {"type": ["string", "null"]}, "capabilities": {"type": "object", "additionalProperties": {"type": "boolean"}}}},
...
      "AssetResult": {"type": "object", "additionalProperties": false, "required": ["asset"], "properties": {"asset": {"$ref": "#/components/schemas/AssetReference"}, "metadata": {"$ref": "#/components/schemas/AssetMetadata"}}},
      "AssetSearchItem": {"type": "object", "additionalProperties": false, "required": ["resource", "entry", "assetId", "metadata"], "properties": {"resource": {"type": "string"}, "entry": {"$ref": "#/components/schemas/Entry"}, "assetId": {"type": ["string", "null"]}, "metadata": {"$ref": "#/components/schemas/AssetMetadata"}}},
      "AssetSearchResult": {"type": "object", "additionalProperties": false, "required": ["items", "total", "offset", "limit", "scanned", "truncated", "facets"], "properties": {"items": {"type": "array", "items": {"$ref": "#/components/schemas/AssetSearchItem"}}, "total": {"type": "integer", "minimum": 0}, "offset": {"type": "integer", "minimum": 0}, "limit": {"type": "integer", "minimum": 1, "maximum": 100}, "scanned": {"type": "integer", "minimum": 0}, "truncated": {"type": "boolean"}, "facets": {"type": "object", "additionalProperties": false, "required": ["resources", "types", "extensions"], "properties": {"resources": {"type": "object", "additionalProperties": {"type": "integer", "minimum": 0}}, "types": {"type": "object", "additionalProperties": {"type": "integer", "minimum": 0}}, "extensions": {"type": "object", "additionalProperties": {"type": "integer", "minimum": 0}}}}}},
      "AssetSearchEnvelope": {"type": "object", "additionalProperties": false, "required": ["success", "data"], "properties": {"success": {"const": true}, "data": {"$ref": "#/components/schemas/AssetSearchResult"}}},
...
      "AltTranslations": {"type": "object", "maxProperties": 20, "propertyNames": {"pattern": "^[a-z]{2,3}(?:-[a-z0-9]{2,8})*$"}, "additionalProperties": {"type": "string", "maxLength": 1000}},
      "AssetMetadata": {"type": "object", "additionalProperties": false, "required": ["alt", "title", "tags", "version", "updatedAt"], "properties": {"alt": {"type": ["string", "null"], "maxLength": 1000}, "altTranslations": {"$ref": "#/components/schemas/AltTranslations"}, "title": {"type": ["string", "null"], "maxLength": 200}, "tags": {"type": "array", "maxItems": 20, "uniqueItems": true, "items": {"type": "string", "maxLength": 50}}, "version": {"type": "integer", "minimum": 1}, "updatedAt": {"type": "integer", "minimum": 0}}},
...
    },
    "responses": {
...
      "Error": {"description": "Stable machine-readable error envelope", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}},
      "AssetSearch": {"description": "Bounded workspace-authorized asset search", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AssetSearchEnvelope"}}}},
...
      ,"Asset": {"description": "Stable editor-facing asset reference", "content": {"application/json": {"schema": {"type": "object", "additionalProperties": false, "required": ["success", "data"], "properties": {"success": {"const": true}, "data": {"$ref": "#/components/schemas/AssetResult"}}}}}}
...
    }
  }
}